Implement DME and LOC tolerance tools with UI and functionality
- Added QPANSOPYDMEToleranceDockWidget for DME tolerance calculations.
- Created QPANSOPYLOCDMEToleranceDockWidget subclass for LOC tolerance.
- Developed dme_tolerance.py module for tolerance geometry calculations.
- Introduced new UI files for DME and LOC tolerance dock widgets.
- Added icons for DME and LOC tolerance tools.
- Updated qt_compat.py to support QToolButton across Qt versions.
- Removed obsolete NDB/DME tolerance UI files.
