+++ Q_Pansopy/metadata.txt
@@ -30,6 +30,7 @@ changelog=
  - Require 'segment' == 'departure' for LNAV SID and rename its output layer to 'PBN RNAV 1/2 Departure'
  - OMNI SID DER marker: plain green triangle plus a connector line to the runway edge when CWY > 0
  - Store Area 1/2/3 distances (meters, 4 decimals) in the OMNI SID output attribute table
+ - Combine VOR/DME and NDB/DME tolerance tools under one CONV toolbar dropdown icon and add LOC/DME Tolerance (2.4° default)
  Version 0.4.0:
  - Fixes to initial realease to code logic including QGIS Plugin Store Compliance
  - PBN 15 NM and 30 NM targets included
